refactor: move my library page to component in teacher views
